feat(raft): pass leader election
This lab was made significantly more difficult by channels in all
honesty. There was a lot of nuance to understanding when it could block
or not block, and how to deal with non-blocking. Also the tricky bug in
ticker with default being non-blocking, when we wanted it to be
blocking. And then there was the invalid state invariant that needed to
be held in transition to follower (I still don't understand that one).
Either way, quite tough, most time was spent just understanding the
paper and not even programming.